pyproject.toml
[tool.poetry]
name = "fastapi_shop_backend"
version = "0.1.0"
description = ""
authors = ["dev <alil.a@mail.ru>"]
readme = "README.md"
packages = [{include = "fastapi_shop_backend"}]
#package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
loguru = "^0.7.2"
redis = "^4.6.0"
fastapi = {extras = ["all"], version = "^0.111.0"}
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
sqlalchemy = "^2.0.30"
asyncpg = "^0.29.0"
cryptography = "^42.0.7"
alembic = "^1.13.1"
gunicorn = "^20.1.0"
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"